requirements/base.txt
# This file has the base requirements needed to run the application. Any
# specific python requirements for developing, testing, releasing are in their
# appropriate *.txt file. Please be mindful of where any requirement is
# necessary.

Django==1.8.2
ipaddress==1.0.7
psycopg2==2.6.1
pygithub3==0.5.1
requests==2.7.0
factory-boy==2.5.2
